+++ csv/parsers/accointing/accointing.go
@@ -8,7 +8,6 @@ import (
 	"github.com/DefiantLabs/cosmos-tax-cli/config"
 	"github.com/DefiantLabs/cosmos-tax-cli/cosmos/modules/bank"
 	"github.com/DefiantLabs/cosmos-tax-cli/cosmos/modules/distribution"
-	"github.com/DefiantLabs/cosmos-tax-cli/cosmos/modules/staking"
 	"github.com/DefiantLabs/cosmos-tax-cli/csv/parsers"
 	"github.com/DefiantLabs/cosmos-tax-cli/db"
 	"github.com/DefiantLabs/cosmos-tax-cli/osmosis/modules/gamm"
@@ -261,9 +260,9 @@ func ParseTx(address string, events []db.TaxableTransaction) ([]parsers.CsvRow, 
 			rows = append(rows, ParseMsgMultiSend(address, event))
 		} else if distribution.IsMsgFundCommunityPool[event.Message.MessageType] {
 			rows = append(rows, ParseMsgFundCommunityPool(address, event))
-		} else if staking.IsMsgWithdrawValidatorCommission[event.Message.MessageType] {
+		} else if distribution.IsMsgWithdrawValidatorCommission[event.Message.MessageType] {
 			rows = append(rows, ParseMsgWithdrawValidatorCommission(address, event))
-		} else if staking.IsMsgWithdrawDelegatorReward[event.Message.MessageType] {
+		} else if distribution.IsMsgWithdrawDelegatorReward[event.Message.MessageType] {
 			rows = append(rows, ParseMsgWithdrawDelegatorReward(address, event))
 		} else if gamm.IsMsgSwapExactAmountIn[event.Message.MessageType] {
 			rows = append(rows, ParseMsgSwapExactAmountIn(address, event))
